requirements.txt
# Core dependencies
ibapi>=10.19.01
pydantic>=2.0.0
pyyaml>=6.0
fastmcp>=2.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Optional: Environment variables support
python-dotenv>=1.0.0
